% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_pass.R
\name{create_pass}
\alias{create_pass}
\title{Creates a password}
\usage{
create_pass(
  nwords = 4,
  min_l = 4,
  max_l = 12,
  sep = "-",
  complex = FALSE,
  seed = NULL
)
}
\arguments{
\item{nwords}{Number of words}

\item{min_l}{Minimum word length}

\item{max_l}{Maximum word length}

\item{sep}{Separator between words}

\item{complex}{Whether to allow the full dictionary set of complex words}

\item{seed}{Set random seed (unset by default)}
}
\description{
Creates a password
}
