% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keyring.r
\name{.normalize_id}
\alias{.normalize_id}
\title{Normalize a \code{.gpg-id} recipient identifier for keyring lookups}
\usage{
.normalize_id(id)
}
\description{
Normalize a \code{.gpg-id} recipient identifier for keyring lookups
}
\keyword{internal}
