% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/key.r
\name{pass_default_key}
\alias{pass_default_key}
\title{Retrieve the session's default key}
\usage{
pass_default_key()
}
\value{
An \code{rpass_key} object.
}
\description{
The key most recently set via \code{pass_key(..., set_default = TRUE)}.
}
\examples{
keys <- rpass:::generate_test_keypair("Demo <demo@example.com>", "hunter2")
secret_path <- tempfile(fileext = ".asc")
writeLines(keys$secret, secret_path)

pass_key(secret_path, passphrase = "hunter2") # set_default = TRUE by default
pass_default_key()
}
