% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keyring.R
\name{pass_list_recipients}
\alias{pass_list_recipients}
\title{List public keys in the local rpass keyring}
\usage{
pass_list_recipients(keyring = pass_keyring_dir())
}
\arguments{
\item{keyring}{Local keyring directory, see \code{\link[=pass_keyring_dir]{pass_keyring_dir()}}.}
}
\value{
A data frame with columns \code{fingerprint} and \code{path}.
}
\description{
List public keys in the local rpass keyring
}
