% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{public_key_fingerprint}
\alias{public_key_fingerprint}
\title{Fingerprint of an armored OpenPGP public key, for matching against \code{.gpg-id}
recipient identifiers.}
\usage{
public_key_fingerprint(armor_text)
}
\description{
Fingerprint of an armored OpenPGP public key, for matching against \code{.gpg-id}
recipient identifiers.
}
