% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list.R
\name{.print_tree_level}
\alias{.print_tree_level}
\title{Recursively render a list of path-segment vectors as a \code{tree}-style listing}
\usage{
.print_tree_level(segments, prefix)
}
\description{
Recursively render a list of path-segment vectors as a \code{tree}-style listing
}
\keyword{internal}
