% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{encrypt_message}
\alias{encrypt_message}
\title{Encrypt plaintext to one or more recipients (armored public keys), producing
a binary OpenPGP message compatible with the \code{.gpg} format used by \code{pass}.}
\usage{
encrypt_message(plaintext, recipient_armors)
}
\description{
Encrypt plaintext to one or more recipients (armored public keys), producing
a binary OpenPGP message compatible with the \code{.gpg} format used by \code{pass}.
}
