% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/key.R
\name{pass_default_key}
\alias{pass_default_key}
\title{Retrieve the session's default key}
\usage{
pass_default_key()
}
\value{
An \code{rpass_key} object.
}
\description{
The key most recently set via \code{pass_key(..., set_default = TRUE)}.
}
