% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.r
\name{public_key_fingerprint}
\alias{public_key_fingerprint}
\title{Fingerprint of an armored OpenPGP public key, for matching against \code{.gpg-id}
recipient identifiers.}
\usage{
public_key_fingerprint(armor_text)
}
\arguments{
\item{armor_text}{An armored OpenPGP public key.}
}
\value{
The key's fingerprint, as an uppercase hex string.
}
\description{
Fingerprint of an armored OpenPGP public key, for matching against \code{.gpg-id}
recipient identifiers.
}
