% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpass-package.r
\docType{package}
\name{rpass-package}
\alias{rpass}
\alias{rpass-package}
\title{rpass: Native Client for 'pass'-Compatible Password Stores}
\description{
Read and manage entries in a password store compatible with 'pass', the standard unix password manager \url{https://www.passwordstore.org/}, without shelling out to the 'pass' or 'gpg' binaries at runtime. OpenPGP parsing, decryption and encryption are implemented natively via the Rust 'pgp' crate, embedded through 'extendr'/'rextendr'.
}
\author{
\strong{Maintainer}: Giuseppe Acito \email{giuseppe.acito@gmail.com}

Authors:
\itemize{
  \item Giuseppe Acito \email{giuseppe.acito@gmail.com}
}

}
\keyword{internal}
